requirements.txt
langchain_community
tiktoken
langchainhub
langchain
langgraph
tavily-python
python-dotenv
google-generativeai
langchain-google-genai
langchain-nomic
langchain-text-splitters
langchain-openai
ipykernel
chromadb 
beautifulsoup4
pymupdf
pydantic
httpx
fastapi
uvicorn
colorama